# Serialization of StoredDocument objects as protobuf