--- a/docs/narrow_down.proto.rst
+++ b/docs/narrow_down.proto.rst
@@ -1,0 +1,21 @@
+narrow\_down.proto package
+==========================
+
+Submodules
+----------
+
+narrow\_down.proto.stored\_document module
+------------------------------------------
+
+.. automodule:: narrow_down.proto.stored_document
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: narrow_down.proto
+   :members:
+   :undoc-members:
+   :show-inheritance: