--- a/narrow_down/data_types.py
+++ b/narrow_down/data_types.py
@@ -8,6 +8,8 @@ from typing import NewType, Optional
 import numpy as np
 from numpy import typing as npt
 
+from narrow_down.proto.stored_document import StoredDocumentProto
+
 
 class TooLowStorageLevel(Exception):
     """Raised if a feature is used for which a higher storage level is needed."""
@@ -49,25 +51,28 @@ class StoredDocument:
     data: Optional[str] = None
     """Payload to persist together with the document in the internal data structures."""
 
-    # TODO: Pickle is far from optimal here.
-    #   Maybe https://github.com/ultrajson/ultrajson is an alternative?
     def serialize(self, storage_level: StorageLevel) -> bytes:
         """Serialize a document to bytes."""
-        return pickle.dumps(
-            dataclasses.asdict(
-                self,
-                dict_factory=lambda items: {
-                    k: v for k, v in items if k in _FIELDS_FOR_STORAGE_LEVEL[storage_level]
-                },
-            )
-        )  # noqa
+        return StoredDocumentProto(
+            fingerprint=list(self.fingerprint)
+            if self.fingerprint is not None and storage_level & StorageLevel.Fingerprint
+            else [],
+            **{f: getattr(self, f) for f in _FIELDS_FOR_STORAGE_LEVEL[storage_level]},
+        ).SerializeToString()
 
     @staticmethod
     def deserialize(doc: bytes, id_: int) -> "StoredDocument":
         """Deserialize a document from bytes."""
-        d = pickle.loads(doc)  # noqa
-        d["id_"] = id_
-        return StoredDocument(**d)
+        p = StoredDocumentProto.FromString(doc)
+        return StoredDocument(
+            id_=id_,
+            document=p.document if p.document else None,
+            exact_part=p.exact_part if p.exact_part else None,
+            fingerprint=Fingerprint(np.array(p.fingerprint, dtype=np.uint32))
+            if p.fingerprint
+            else None,
+            data=p.data if p.data else None,
+        )
 
     def without(self, *attributes: str) -> "StoredDocument":
         """Create a copy with the specified attributes left out.
@@ -87,7 +92,7 @@ class StoredDocument:
 _FIELDS_FOR_STORAGE_LEVEL = {
     StorageLevel.Minimal: {"data"},
     StorageLevel.Document: {"data", "document"},
-    StorageLevel.Fingerprint: {"data", "exact_part", "fingerprint"},
-    StorageLevel.Full: {"data", "document", "exact_part", "fingerprint"},
+    StorageLevel.Fingerprint: {"data", "exact_part"},
+    StorageLevel.Full: {"data", "document", "exact_part"},
 }
 """Fields of StoredDocument which need to be serialized to reach a certain storage level."""