--- a/pyproject.toml
+++ b/pyproject.toml
@@ -61,6 +61,7 @@ dev = [
     "safety~=1.10",
     "nox",
     "mypy",
+    "mypy-protobuf",
     "sphinx~=4.2",
     "bump2version~=1.0",
     "pytest~=6.2",
@@ -72,6 +73,7 @@ dev = [
     "watchdog[watchmedo]~=2.1",
     "myst-parser~=0.16",
     "flake8-pylint~=0.1",
+    "protoc-wheel-0"
 ]
 
 [tool.coverage.paths]
@@ -196,8 +198,8 @@ disable = [
 max-args = 10  # Maximum number of arguments for functions
 max-attributes = 15  # Maximum number of attributes for a class
 
-[tool.mypy]
-exclude = '^narrow_down/proto/.+_pb2.py$'  # Ignore generated protobufs
+#[tool.mypy]
+#exclude = '^narrow_down/proto/.+_pb2.py$'  # Ignore generated protobufs
 
 [[tool.mypy.overrides]]
 module = [