--- a/pyproject.toml
+++ b/pyproject.toml
@@ -198,9 +198,6 @@ disable = [
 max-args = 10  # Maximum number of arguments for functions
 max-attributes = 15  # Maximum number of attributes for a class
 
-#[tool.mypy]
-#exclude = '^narrow_down/proto/.+_pb2.py$'  # Ignore generated protobufs
-
 [[tool.mypy.overrides]]
 module = [
     "pytest",