--- a/tests/test_data_types.py
+++ b/tests/test_data_types.py
@@ -7,6 +7,11 @@ import pytest
 from narrow_down.data_types import Fingerprint, StorageLevel, StoredDocument
 
 
+@pytest.mark.parametrize("data", [None, "user data"])
+@pytest.mark.parametrize("fingerprint", [None, Fingerprint(np.array([1]))])
+@pytest.mark.parametrize("exact_part", [None, "text_exact_part"])
+@pytest.mark.parametrize("document", [None, "text_document"])
+@pytest.mark.parametrize("id_", [None, 0, 1])
 @pytest.mark.parametrize(
     "storage_level, expected_fields",
     [
@@ -16,13 +21,15 @@ from narrow_down.data_types import Fingerprint, StorageLevel, StoredDocument
         (StorageLevel.Full, {"data", "document", "exact_part", "fingerprint"}),
     ],
 )
-def test_stored_document_serialization(storage_level, expected_fields):
+def test_stored_document_serialization(
+    storage_level, expected_fields, id_, document, exact_part, fingerprint, data
+):
     document = StoredDocument(
-        id_=5,
-        document="abcd",
-        exact_part="exact:part",
-        fingerprint=Fingerprint(np.array([1])),
-        data="user data",
+        id_=id_,
+        document=document,
+        exact_part=exact_part,
+        fingerprint=fingerprint,
+        data=data,
     )
     deserialized = StoredDocument.deserialize(document.serialize(storage_level), id_=None)
     for field in dataclasses.fields(deserialized):