… Cargo.lock
… narrow_down/proto/stored_document.py
… narrow_down/proto/stored_document_pb2.py
… narrow_down/proto/stored_document_pb2.pyi