Serializing StoredDocument objects with protobuf instead of with pickle to save memory / storage capacity. As bonus it is also faster than pickle.

Test with:
```
        storage_level=nd.data_types.StorageLevel.Minimal,
        similarity_threshold=0.8,
        max_false_negative_proba=0.05,
        max_false_positive_proba=0.05,
        tokenize="char_ngrams(3)",
```
which gives:
```
MinhashLshConfig(n_hashes=64, n_bands=6, rows_per_band=10)
```

Test code:
<details>

```python
import time

import pandas as pd
from pympler.asizeof import asizeof
from humanize.filesize import naturalsize

import narrow_down as nd

def preprocess(df):
    return df.assign(
        normalized=df["Summary"].astype(str).str.lower().str.replace(r"[^a-z]", "")
    )


df = pd.read_csv("data/AmazonFineFoodReviews/Reviews.csv", dtype=str)[
    ["Summary"]
].dropna()
df = preprocess(df)

metrics = dict(index_time_s=[], memory_bytes=[], query_time_s=[])
sizes = (100, 1000, 10000, 100000)
threshold = 0.8

for n_items in sizes:

    start = time.perf_counter()

    simstore = nd.similarity_store.SimilarityStore(
        storage_level=nd.data_types.StorageLevel.Full,
        similarity_threshold=threshold,
        max_false_negative_proba=0.05,
        max_false_positive_proba=0.05,
        tokenize="char_ngrams(3)",
    )
    print(simstore._lsh_config)
    await simstore.initialize()
    for docid, document in df.normalized[:n_items].iteritems():
        await simstore.insert(document, document_id=docid)

    #     print(len(train), "items indexed")
    metrics["index_time_s"].append(round(time.perf_counter() - start, 3))
    metrics["memory_bytes"].append(naturalsize(asizeof(simstore)))

    times = 0
    for s in df.normalized[-1000:]:
        start = time.perf_counter()
        await simstore.query(s)
        times += time.perf_counter() - start
    metrics["query_time_s"].append(round(times / 1000, 5))

pd.DataFrame(metrics, index=sizes)
naturalsize(asizeof(simstore._storage._buckets))
naturalsize(asizeof(simstore._storage._documents))
```

</details>

## Results for StorageLevel = Minimal

### Indexing before (with pickle):
| document_count | index_time_s | memory_bytes | query_time_s |
|----------------|--------------|--------------|--------------|
|       100      | 0.007        | 185.6 kB     | 0.00003      |
|      1000      | 0.052        | 1.7 MB       | 0.00003      |
|      10000     | 0.491        | 14.9 MB      | 0.00004      |
|     100000     | 5.150        | 133.3 MB     | 0.00019      |

For 100,000:
Size of simstore._storage._buckets: 122.9 MB
Size of simstore._storage._documents: 14.0 MB

### Indexing after (with protobuf):
| document_count | index_time_s | memory_bytes | query_time_s |
|----------------|--------------|--------------|--------------|
|       100      | 0.005        | 180.0 kB     | 0.00004      |
|      1000      | 0.038        | 1.7 MB       | 0.00003      |
|      10000     | 0.348        | 14.3 MB      | 0.00005      |
|     100000     | 4.102        | 127.7 MB     | 0.00024      |

For 100,000:
Size of simstore._storage._buckets: 122.9 MB
Size of simstore._storage._documents: 8.4 MB

## Results for StorageLevel = Document

### Before:
| document_count | index_time_s | memory_bytes | query_time_s |
|----------------|--------------|--------------|--------------|
|       100      | 0.017        | 189.2 kB     | 0.00036      |
|      1000      | 0.055        | 1.7 MB       | 0.00003      |
|      10000     | 0.511        | 15.3 MB      | 0.00005      |
|     100000     | 5.669        | 137.1 MB     | 0.00020      |

Size of simstore._storage._documents: 17.8 MB

### After:
| document_count | index_time_s | memory_bytes | query_time_s |
|----------------|--------------|--------------|--------------|
|       100      | 0.005        | 185.7 kB     | 0.00004      |
|      1000      | 0.039        | 1.7 MB       | 0.00003      |
|      10000     | 0.369        | 14.9 MB      | 0.00015      |
|     100000     | 4.185        | 133.5 MB     | 0.00027      |

Size of simstore._storage._documents: 14.2 MB

## Results for StorageLevel = Full

### Before:
| document_count | index_time_s | memory_bytes | query_time_s |
|----------------|--------------|--------------|--------------|
|       100      | 0.022        | 231.5 kB     | 0.00011      |
|      1000      | 0.065        | 2.2 MB       | 0.00003      |
|      10000     | 0.639        | 19.5 MB      | 0.00009      |
|     100000     | 6.785        | 179.3 MB     | 0.00054      |

Size of simstore._storage._documents: 60.0 MB

### After:
| document_count | index_time_s | memory_bytes | query_time_s |
|----------------|--------------|--------------|--------------|
|       100      | 0.009        | 213.3 kB     | 0.00011      |
|      1000      | 0.048        | 2.0 MB       | 0.00004      |
|      10000     | 0.584        | 17.7 MB      | 0.00013      |
|     100000     | 5.291        | 161.2 MB     | 0.00087      |

Size of simstore._storage._documents: 41.9 MB